avito_service/requirements.txt
beautifulsoup4==4.8.2
bs4==0.0.1
lxml==4.5.0
requests==2.22.0
urllib3==1.25.8
psycopg2==2.8.5
peewee==3.13.3
python-telegram-bot==12.7
python-dotenv==0.13.0
PySocks==1.7.1